Restructure Python packages to src/package/{cli,gui,core}.py layout
- mukill, mufile, muexpression, muspot: move to src/package/ structure
- Add PyTorch cu130 index to muspot (matching mukill)
- Root pyproject: add workspace deps, dev group (pytest), sources
- mufile tests: update import to mufile.core
