--- muexpression/gui/__init__.py
@@ -1,5 +1,0 @@
-"""muexpression GUI – CustomTkinter interface for analyze and plot."""
-
-from gui.app import main  # noqa: F401
-
-__all__ = ["main"]
