--- mufile/gui/__init__.py
@@ -1,5 +1,0 @@
-"""mufile GUI – CustomTkinter interface for crop, convert, movie."""
-
-from gui.app import main  # noqa: F401
-
-__all__ = ["main"]
