+++ mufile/src/mufile/__init__.py
@@ -1,0 +1,1 @@
+"""mufile – microscopy file utilities: crop TIFFs into zarr, convert ND2 to TIFF."""
