--- mukill/gui/__init__.py
@@ -1,5 +1,0 @@
-"""mukill GUI – CustomTkinter interface for dataset, train, predict, plot, clean."""
-
-from gui.app import main  # noqa: F401
-
-__all__ = ["main"]
