+++ mukill/src/mukill/__init__.py
@@ -1,0 +1,1 @@
+"""mukill – train classifiers and analyze kill curves for micropattern experiments."""
