--- muspot/gui/__init__.py
@@ -1,5 +1,0 @@
-"""muspot GUI – CustomTkinter interface for detect and plot."""
-
-from gui.app import main  # noqa: F401
-
-__all__ = ["main"]
