+++ muspot/src/muspot/__init__.py
@@ -1,0 +1,1 @@
+"""muspot – detect fluorescent spots in micropattern crops using spotiflow."""
